refactor out sql prompts
