+++ packages/database/src/sql/mysql.ts
@@ -2,10 +2,7 @@ import { config } from '../config'
 import { createConnection } from 'mysql2/promise'
 import { join } from 'path'
 import { exec } from '../shell'
-import input from '@inquirer/input'
-import inquirer from 'inquirer'
-import fs from 'fs'
-import { BaseSql } from './base'
+import { BaseSql, promptDump, promptRestore } from './base'
 
 const noDatabaseConfig = {
   host: config.host,
@@ -93,45 +90,17 @@ export class Mysql extends BaseSql {
    * @param {string} database
    */
   restorePrompted(database: string) {
-    type Answers = {
-      sql: string
-      confirm: 'yes' | 'no'
-    }
-    const sqlDir = join(config.rootDir, '.sql')
-    const sqlList = fs.readdirSync(sqlDir).filter((x) => x.endsWith('.sql'))
-    inquirer
-      .prompt([
-        {
-          type: 'list',
-          name: 'sql',
-          message: 'Restore from .sql file?',
-          choices: sqlList,
-        },
-        {
-          type: 'list',
-          name: 'confirm',
-          message: `Confirm overwrite database [${database}]?`,
-          choices: ['yes', 'no'],
-        },
-      ])
-      .then(async (answers: Answers) => {
-        if (answers.confirm === 'no') {
-          console.log('cancelled.')
-          return
-        }
-        await this.restoreFromFile(database, answers.sql)
-      })
+    promptRestore(database).then(async (answers) => {
+      if (answers.confirm === 'no') {
+        console.log('cancelled.')
+        return
+      }
+      await this.restoreFromFile(database, answers.sql)
+    })
   }
 
-  /**
-   * dump a database snapshot to an .sql file
-   * @param {string} database
-   */
   async dump(database: string) {
-    const filename = await input({
-      message: 'Enter filename (without .sql):',
-      default: 'backup',
-    })
+    const filename = await promptDump()
     const withExt = filename.concat('.sql')
     const file = join(config.rootDir, '.sql', withExt)
     const u = config.username ? `-u ${config.username}` : ''
